% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHMcatalog.R
\name{CHMcatalog}
\alias{CHMcatalog}
\title{Return a catalog of DESC chemistry data}
\usage{
CHMcatalog(channel, n = 10)
}
\arguments{
\item{channel}{connection handle as returned by odbcConnect}

\item{n}{minimum sample count}
}
\value{
A data frame
}
\description{
The function CHMcatalog can be used to return a catalog of chemistry data in
the DESC database. There are only two argumnets. One to define the connection
handle and the other to limit the size of the catalog by sample count. Many
sites within the DESC chemistry database have only 1 or two samples and are
excluded by default. This can be changed by changing 'n' which defaults to 10
but can be set to any value.
}
\examples{
# Establish connection with database
CHM = odbcConnect("ORA", uid = "chrism", pwd = "chrism_pw")

# Grab a catalog of available chemistry data
Catalog = CHMcatalog(CHM)

# Grab a catalog of all available chemistry data
Catalog = CHMcatalog(CHM, n = 0)

}

