% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHMselect.R
\name{CHMselect}
\alias{CHMselect}
\title{Query DESC chemistry database}
\usage{
CHMselect(channel, STATION = NULL, LDESC = NULL, TEST_NAME = NULL,
  SUBNUM = NULL, startDate = "1970-01-1 00:00:00", endDate = Sys.Date())
}
\arguments{
\item{channel}{connection handle as returned by odbcConnect}

\item{STATION}{station number}

\item{LDESC}{location description given during sample submissions}

\item{TEST_NAME}{the test name short form found in the lab submission manual}

\item{SUBNUM}{submission number}

\item{startDate}{access data from this date forward}

\item{endDate}{access data up to this date}
}
\value{
A data frame
}
\description{
The function CHMselect can be used to get data from the DESC chemistry database.
A standard SQL query is issued to the database and the values are
returned in dataframe. The data returned can be limited by a start
and/or end date defined by the user.
}
\examples{
# Establish connection with database
CHM = odbcConnect("ORA", uid = "chrism", pwd = "chrism_pw")

# Extract all data by STATION
tmp <- CHMselect(CHM, STATION = '03008557402')

# Extract data by STATION and multiple TEST_NAME.
HP4 = CHMselect(CHM, STATION = '03008557402', TEST_NAME = c('PPUT1','PPUT2'))

# Limit extraction by date.
tmp = CHMselect(CHM, STATION = '03008557402', TEST_NAME = 'PPUT1',
 startDate = "2015-09-01", endDate = "2015-11-01")

}

